screen structure implemented
linked screens with link for user path
