+++ internal/client/screens/auth_register.go
